// src/app/models/res-aprendizaje.ts
export class ResAprendizaje {
    id_resultado: string;
    competencia: string;
    tipo_saber: string;
    saber_asociado: string;
    taxonomia: string;
    dominio_tratar: string;
    nivel_dominio: string;
    verbo: string;
    resultado_ap: string;

    constructor() {
        this.id_resultado = '';
        this.competencia = '';
        this.tipo_saber = '';
        this.saber_asociado = '';
        this.taxonomia = '';
        this.dominio_tratar = '';
        this.nivel_dominio = '';
        this.verbo = '';
        this.resultado_ap = '';
    }
}